pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "icongen"
version = "1.0.0"
description = "CLI tool to generate icons at multiple sizes from a logo image"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
]

[project.scripts]
icongen = "icongen.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["icongen*"]
